[PATCH] USB: ohci D3 resume fix
This fixes a problem that cropped up resuming OHCI from PCI D3 on
NForce2.  Evidently the register controlling frame timing gets
clobbered in D3, but not other registers ... a "by the book" reinit
seems to solve this particular problem.  (And ought to help a few
startup glitches on other implementations.  Linux never used that
toggle bit before, and has had to struggle with glitchy init...)

It also updates some diagnostics to be bit more useful, and tries
to avoid a particular "wakeup while suspending" glitch that seems
particular to one particular type of Australian mouse.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>