[PATCH] S2io: Changed copyright and added support for Xframe II
1. Change of Copyright to reflect S2io's new name "Neterion Inc."
2. Updated driver version number.
3. Add an additional PCI device id to support Neterion's new Xframe II.

Some background info on Xframe-II, just in case -

The NIC was announced back in January (see
http://www.serverwatch.com/news/article.php/3464871)

Xframe II is a PCI-X 2.0 DDR adapter designed to work in PCI-X 2.0 servers
(it is also backwards compatible with current pci-x and pci slots); it's a
first card to overcome pci-x 133 throughput bottleneck. Some of these pci-x
2.0 servers are available now and more are coming later in the year.

Xframe II is completely backward compatible and hence the current driver
will work as-is, except the device id change.

There are some additional features/stateless offloads in Xframe-II
as well; we plan to submit patches for these soon.

Signed-off-by: Ravinandan Arakali <ravinandan.arakali@neterion.com>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>