[PATCH] nfsd4: acl don't set named attrs
We were expecting {read,write}_named_attributes to be set in sync with the
read and write bits.  But we don't actually support named attributes, and when
we eventually do, it's unlikely that we'll want to use these ACL mode bits to
control access to them.  So let's just stop supporting these two bits.

This also updates our mapping to the latest ietf draft.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>