[PATCH] I2C: Kill outdated defines in i2c.h
Some defines in i2c.h (I2C_CLIENT_MODPARM and friends) are now useless.
They should have been removed when the i2c client parameters were
converted from MODULE_PARAM to module_parm_array, but where not. This
patch removes them now.

Additionally, it moves the definition of I2C_CLIENT_MAX_OPTS next to
where it is used rather than 220 lines before, which is preferable IMHO.

As a side note, I think that there is a bug in the way these options are
handled. The i2c code looks for I2C_CLIENT_END as a list terminator, but
if the maximum number of parameters are actually provided, no terminator
will be left. It's rather unlikely to happen because nobody will
probably ever provide that many parameters, but this should probably be
fixed. I'll address this issue later, since I plan to completely rewrite
the way these parameters are handled anyway.

Signed-off-by: Jean Delvare <khali@linux-fr.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>