[PATCH] Fix Oops in MXB driver (v4l2 subsystem)
This fixes a NULL pointer dereference Oops in my "Multimedia eXtension
Board" driver. 

The tda9840 i2c driver dereferences the argument pointer, but the MXB
driver is supplying a NULL pointer for one of the commands.  The patch
makes this one command behave like the others, ie.  it expects an int
argument.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>